+++ src/popup/components/PingDotsLoader/PingDotsLoader.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './ping-dots-loader.pcss';
+
+export const PingDotsLoader = () => (
+    <div className="ping-dots-loader">
+        <span className="ping-dots-loader__dot" />
+        <span className="ping-dots-loader__dot" />
+        <span className="ping-dots-loader__dot" />
+    </div>
+);