+++ src/popup/components/ui/SkeletonEndpoint/SkeletonEndpoint.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import './skeleton-endpoint.pcss';
+
+export const SkeletonEndpoint = () => {
+    return (
+        <div className="skeleton-endpoint">
+            <div className="skeleton-endpoint__empty skeleton-endpoint__empty-flag" />
+            <div className="skeleton-endpoint__container">
+                <div className="skeleton-endpoint__empty skeleton-endpoint__empty-title" />
+                <div className="skeleton-endpoint__empty skeleton-endpoint__empty-desc" />
+            </div>
+        </div>
+    );
+};