+++ src/popup/components/ui/SkeletonFooter/index.ts
@@ -1,1 +1,0 @@
-export { SkeletonFooter } from './SkeletonFooter';